requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
pdf2docx==0.5.7
docx2pdf==0.1.8
python-pptx==0.6.21
Werkzeug==2.3.7
